Added custom model and animation format. Removed assimp dependency from Vita builds
Models are now only loaded with Assimp when importing them into the Editor. They are then saved in a custom format which is what is loaded by the engine. This will probably allow for faster load times and it also allowed to remove the assimp dependency from PS Vita builds. Also fixed some memory leaks