Fixed #1
A default material is created. If it fails to be created the program exits. If a shader fails to compile we don't load the material and return the default one which might not even show anything on the screen.  Some validation messages might show too, but it's better than it crashing and it's only temporary until we fix the main material.